Inline aggregate override
