+++ graphrag/index/flows/create_base_text_units.py
@@ -3,13 +3,18 @@
 
 """All the steps to transform base text_units."""
 
+from dataclasses import dataclass
 from typing import Any, cast
 
 import pandas as pd
-from datashaper import VerbCallbacks
+from datashaper import (
+    FieldAggregateOperation,
+    Progress,
+    VerbCallbacks,
+    aggregate_operation_mapping,
+)
 
 from graphrag.index.utils import gen_md5_hash
-from graphrag.index.verbs.overrides.aggregate import aggregate_df
 from graphrag.index.verbs.text.chunk.text_chunk import chunk_df
 
 
@@ -28,7 +33,9 @@ def create_base_text_units(
         zip(*[sort[col] for col in ["id", "text"]], strict=True)
     )
 
-    aggregated = aggregate_df(
+    callbacks.progress(Progress(percent=0))
+
+    aggregated = _aggregate_df(
         sort,
         groupby=[*chunk_by_columns] if len(chunk_by_columns) > 0 else None,
         aggregations=[
@@ -39,6 +46,8 @@ def create_base_text_units(
             }
         ],
     )
+
+    callbacks.progress(Progress(percent=1))
 
     chunked = chunk_df(
         aggregated,
@@ -67,3 +76,57 @@ def create_base_text_units(
     return cast(
         pd.DataFrame, chunked[chunked[chunk_column_name].notna()].reset_index(drop=True)
     )
+
+
+# TODO: would be nice to inline this completely in the main method with pandas
+def _aggregate_df(
+    input: pd.DataFrame,
+    aggregations: list[dict[str, Any]],
+    groupby: list[str] | None = None,
+) -> pd.DataFrame:
+    """Aggregate method definition."""
+    aggregations_to_apply = _load_aggregations(aggregations)
+    df_aggregations = {
+        agg.column: _get_pandas_agg_operation(agg)
+        for agg in aggregations_to_apply.values()
+    }
+    if groupby is None:
+        output_grouped = input.groupby(lambda _x: True)
+    else:
+        output_grouped = input.groupby(groupby, sort=False)
+    output = cast(pd.DataFrame, output_grouped.agg(df_aggregations))
+    output.rename(
+        columns={agg.column: agg.to for agg in aggregations_to_apply.values()},
+        inplace=True,
+    )
+    output.columns = [agg.to for agg in aggregations_to_apply.values()]
+    return output.reset_index()
+
+
+@dataclass
+class Aggregation:
+    """Aggregation class method definition."""
+
+    column: str | None
+    operation: str
+    to: str
+
+    # Only useful for the concat operation
+    separator: str | None = None
+
+
+def _get_pandas_agg_operation(agg: Aggregation) -> Any:
+    if agg.operation == "string_concat":
+        return (agg.separator or ",").join
+    return aggregate_operation_mapping[FieldAggregateOperation(agg.operation)]
+
+
+def _load_aggregations(
+    aggregations: list[dict[str, Any]],
+) -> dict[str, Aggregation]:
+    return {
+        aggregation["column"]: Aggregation(
+            aggregation["column"], aggregation["operation"], aggregation["to"]
+        )
+        for aggregation in aggregations
+    }
