--- graphrag/index/verbs/overrides/__init__.py
@@ -1,8 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""The Indexing Engine overrides package root."""
-
-from .aggregate import aggregate
-
-__all__ = ["aggregate"]
